Tooling: pre-commit hooks, dependency bounds, py.typed, ruff linter, mypy config

+++ quantarhei/qm/hilbertspace/statevector.py
@@ -1,17 +1,16 @@
-# -*- coding: utf-8 -*-
-"""
-    Quantarhei package (http://www.github.com/quantarhei)
+"""Quantarhei package (http://www.github.com/quantarhei)
 
-    statevector module
+statevector module
 
 
 """
 
 import numpy
 
-from ...utils.types import BasisManagedComplexArray
+from ... import COMPLEX
 from ...core.managers import BasisManaged
-from ... import REAL, COMPLEX
+from ...utils.types import BasisManagedComplexArray
+
 
 class StateVector(BasisManaged):
     """Represents a quantum mechanical state vector
@@ -23,13 +22,12 @@ class StateVector(BasisManaged):
 
     Examples
     --------
-
     >>> psi = StateVector(2)
     >>> print(psi.dim)
     2
 
 
-    >>> vec = numpy.zeros((1,3), dtype=REAL)
+    >>> vec = numpy.zeros((1,3), dtype=float)
     >>> psi = StateVector(data=vec)
     Traceback (most recent call last):
     ...
@@ -39,7 +37,7 @@ class StateVector(BasisManaged):
 
     """
 
-    data = BasisManagedComplexArray("data")   
+    data = BasisManagedComplexArray("data")
 
     def __init__(self, dim=None, data=None):
 
@@ -77,7 +75,6 @@ class StateVector(BasisManaged):
         """Scalar product of two StateVectors
 
         """
-
         return numpy.dot(self.data, vec.data)
 
     def norm(self):
@@ -89,44 +86,43 @@ class StateVector(BasisManaged):
 
     def transform(self, SS, inv=None):
         """Transformation of the operator by a given matrix
-        
-        
+
+
         This function transforms the Operator into a different basis, using
         a given transformation matrix.
-        
+
         Parameters
         ----------
-        
         SS : matrix, numpy.ndarray
             transformation matrix
-            
+
         inv : matrix, numpy.ndarray
             inverse of the transformation matrix
-            
-        """        
+
+        """
         if inv is None:
             S1 = numpy.linalg.inv(SS)
         else:
             S1 = inv
 
         self._data = numpy.dot(S1,self._data)
-        
+
 
     def get_DensityMatrix(self):
         """Constructs DensityMatrix from the present StateVector
-        
+
         """
         from .operators import DensityMatrix
 
         rho = DensityMatrix(dim=self.dim)
-        
+
         for ii in range(self.dim):
             for jj in range(self.dim):
                 rho.data[ii,jj] = self.data[ii]*numpy.conj(self.data[jj])
-        
+
         return rho
-        
-        
+
+
     def __str__(self):
         out  = "\nquantarhei.StateVector object"
         out += "\n============================="
